refactor(web): standardize route module folders
