+++ apps/hyperlocalise-web/src/api/app.ts
@@ -12,12 +12,12 @@ import { handleUnexpectedError, notFoundHandler } from "./errors";
 import { createAgentEmailRoutes } from "./routes/agent-email/agent-email.route";
 import { createAgentSlackRoutes } from "./routes/agent-slack/agent-slack.route";
 import { createApiKeyRoutes } from "./routes/api-key/api-key.route";
-import { authRoutes } from "./routes/auth";
+import { authRoutes } from "./routes/auth/auth.route";
 import { createChatRequestRoutes } from "./routes/chat-request/chat-request.route";
 import { createConversationRoutes } from "./routes/conversation/conversation.route";
 import { createGlossaryRoutes } from "./routes/glossary/glossary.route";
 import { createGithubInstallationRoutes } from "./routes/github-installation/github-installation.route";
-import { createGithubWebhookRoutes } from "./routes/github-webhook";
+import { createGithubWebhookRoutes } from "./routes/github-webhook/github-webhook.route";
 import { healthRoutes } from "./routes/health";
 import { createMcpRoutes } from "./routes/mcp/mcp.route";
 import { createWorkspaceJobRoutes } from "./routes/project/job.route";
@@ -25,12 +25,12 @@ import { createProjectRoutes } from "./routes/project/project.route";
 import { createProviderCredentialRoutes } from "./routes/provider-credential/provider-credential.route";
 import { createPublicFileRoutes } from "./routes/public-files/public-files.route";
 import { createPublicJobRoutes } from "./routes/public-jobs/public-jobs.route";
-import { createResendWebhookRoutes } from "./routes/resend-webhook";
-import { createSlackOAuthRoutes } from "./routes/slack-oauth";
-import { createSlackWebhookRoutes } from "./routes/slack-webhook";
+import { createResendWebhookRoutes } from "./routes/resend-webhook/resend-webhook.route";
+import { createSlackOAuthRoutes } from "./routes/slack-oauth/slack-oauth.route";
+import { createSlackWebhookRoutes } from "./routes/slack-webhook/slack-webhook.route";
 import { createFileRoutes } from "./routes/file/file.route";
 import { createTeamRoutes } from "./routes/team/team.route";
-import { workosWebhookRoutes } from "./routes/workos-webhook";
+import { workosWebhookRoutes } from "./routes/workos-webhook/workos-webhook.route";
 import { createTranslationJobEventQueue } from "@/workflows/adapters";
 
 type CreateAppOptions = {
